+++ lending/src/sdk/models/shared/accountingaddress.ts
@@ -6,7 +6,7 @@ import { SpeakeasyBase, SpeakeasyMetadata } from "../../../internal/utils";
 import { AccountingAddressType } from "./accountingaddresstype";
 import { Expose } from "class-transformer";
 
-export class Addressesitems extends SpeakeasyBase {
+export class AccountingAddress extends SpeakeasyBase {
     /**
      * City of the customer address.
      */